gvnfmadapter/requirements.txt
# rest framework
Django==1.11.9
djangorestframework==3.7.7

# for call rest api
httplib2==0.9.2

# for unit test
coverage==4.2
mock==2.0.0
unittest_xml_reporting==1.12.0

# for auto swagger
drf-yasg>=1.2.2

# for the validation feature
flex>=6.11.1
swagger-spec-validator>=2.1.0
